mmu: vma searching functions
In issue #635, we discovered that the way we search for a range of
addresses in our VMA list does not work correctly in newer versions
of Boost. Before we fix this bug, let's move the VMA searching code
which was duplicated several times in mmu.cc to two functions,
find_intersecting_vmas (to search for a range of addresses) and
find_intersecting_vma (to search for a single address).

This patch should make no functional changes to the code - it just
reorganizes it to pave the way for the next patch.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>