mmu: fix bug on newer versions of Boost
As reported in issue #635, while our VMA searching code worked on older
versions of Boost (such as 1.55), on new versions (such as 1.57) it no
longer works correctly, resulting, for example,  in Java crashing.

It turns out that our nice trick of using the alternative comparator
addr_compare to find intersecting vmas is considered illegal by the
Boost developers (see https://svn.boost.org/trac/boost/ticket/11701),
and results in undefined behavior which happened to be what we expected
in Boost 1.55 - but no longer is.

So this patch replaces the nice (but not working) addr_compare trick
for finding the intersecting vmas, with more eleborate (but working) code
which finds separately the beginning and the end of the intersection
range, with a comparator which is compatible with vma_compare.

This code is slightly slower than the previous code because lower_bound
doesn't have a hint where to start, but still it is logarithmic in size
of vma_list, so I believe it is an acceptable implementation. And of
course, without this patch, OSv doesn't work on Fedora 22 or any other
distro which switched to Boost 1.57, so we need this patch :-)

Fixes #635.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>