libc: implement __asprintf_chk
Implement __asprintf_chk(), which the Linux ABI uses instead of asprintf()
when code is compiled with -D_FORTIFY_SOURCE=2.

We do not implement any additional checking compared to the regular
asprintf, but at least it prevents code compiled this way from crashing
on a missing symbol.

Thanks to Justin Činkelj for noticing this function missing, and
proposing a patch.

Fixes #657

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>