build: eliminate warnings from imgedit.py
imgedit.py, used by the Makefile and scripts/build, uses "qemu-nbd" to
modify images - which can be in raw format or other formats (e.g., qcow2).

Recent versions of qemu started to print annoying warnings qemu-nbd needs
to guess the image type of a raw image. So for raw images, we need to
explicitly tell qemu-nbd that they are raw. Instead of adding new options
to imgedit.py for this esoteric case, I decided just to pass the filename
"-f raw file.raw" to imgedit.py - which uses it in an qemu-nbd command
line. Pretty ugly, but it works.

This patch makes progress for issue #663, but some other warnings remain
as well.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>