tests: make tst-rename runnable in Linux too
It is useful that OSv tests can also be compiled for Linux, so we can
confirm that the test is really testing for the correct behavior.

tst-rename had some unnecessary dependency on <osv/debug.h>, which this
patch removes. Additionally, some of the tests try to move a file between
directories (such as "." and "/tmp") which on most Linux distributions
are on different devices so move is disallowed, so we need to also allow
the EXDEV error code.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>