fs: fix rename() permission checking
It is arguable (see issue #593) whether in OSv we should do file permission
checking at all, given that on Linux root user doesn't do it. However, if we
do have file permission checking, let's at least not botch it up!

For rename() we checked if the file being renamed is writable. This is
incorrect - it is perfectly fine to rename an unwritable file. What should
be checked for writeability is file's parent directory - as well as the
destination directory.

This patch also includes a test for renaming an unwritable file, which
succeeds on Linux but failed on OSv, and succeeds after this patch.

This patch also might solve issue #661, although there might be yet another
bug exposed in that issue (where stat() and vn_access() see different
permissions).

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>