fs: fix reference leak in unlink() error
The previous patch to unlink() introduced a subtle bug, where we forget
a vn_unlock() when the parent directory is not writable. This patch adds
it.

The real way to fix this is to use RAII instead of all this explicit (and
error-prone) lock/unlock and gotos, but that would require a substantial
overhaul of this file, and we can do that another time...

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>