run.py: by default, run last built image
It is easy to forget to run "run.py -d" after compiling with "build
mode=debug". Without this "-d" option, the default was to run the release
mode build - even if we just built the debug version.

With this patch, run.py defaults to running the *last* built image,
whether it was a debug or release build. If you have compiled both
versions, and want to explicitly choose the version to run (instead
of running the one you compiled last), you can use the existing "-d"
option to choose the debug build, or the new "-r" option to choose
the release build.

Additionally, this patch add a clearer error message when the requested
image does not exist. For example,

$ scripts/run.py -d
Exception: Image file /home/nyh/osv/build/debug/usr.img does not exist.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>