scripts/build: add slash on prefix
cpiod.so blindly concatenates the prefix we give it with the paths it gets.
So if the prefix doesn't end in a slash (we had "/zfs/zfs" before this patch)
and the path doesn't start with a slash (we had such a case in the ruby app),
the concetenation results in nonsense.

The solution is to end the prefix with a slash: /zfs/zfs/. Relative paths
will be considered, just as before, relative to the zfs root directory.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@scylladb.com>