getrlimit(): abort warning shouldn't be hidden
It is a mistake to call debug() (or kprintf()) and then abort(), because
the debug message doesn't always appear (only if OSv is run with "--verbose")
and then the user will see a bare, unexplained, "Aborted" message.

The better idiom is to call abort(message) - which ensures the message is
printed unconditionally, replacing the usual "Aborted" message from abort().

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Signed-off-by: Pekka Enberg <penberg@scylladb.com>