Makefile: better warning on missing submodules
If someone tries to "make" without running "git submodule update --init"
first, the result are really hard to understand compilation errors.
The README tells the user to run this "git submodule" command, but it is
hidden among a lot of other text and easy to miss.

So this patch adds to the Makefile tests for the important submodules
(those needed by a default run of "make") and if one is missing, tell the
user exactly what's wrong and what to run.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>