sched/elf: Pave the way for creating new namespace
In order to create a new partial namespace a new instance
of elf::program is needed.

This patch keep elf::s_program for all running thread while
creating a way to add a new per application elf::program when
needed.

This idea was suggested by Glauber as an alternative to the
previous series which was somewhat intrusive.

This patch also introduces an application method that allows a new program to be
created for the calling thread. The new program is created at a different slice
of the address space. This effectively imposes a lower amount of memory each
program can use. But the limit is still big enough, and should be enough for
anybody (tm).

Signed-off-by: Benoît Canet <benoit@scylladb.com>
Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>