app: do not ignore running out of ELF namespaces
The current ELF namespace code silently ignores the case of failing to
create a new "program" (ELF namespace). In that case, the program to be
run will not be run on its own namespace, and the caller won't even know,
which can lead to unexpected results.

This situation is not far-fetched because the current
application::new_program() code is limited to 32 namespaces, and even worse,
the current code does not free the namespace id in the elf::program or
osv::application destructor, so even running 33 different programs one
after another (not in parallel) will lead to the same failure.

So as a quick fix, this patch abort()s when we reach this problem.
Of course a better long-term fix would be to do all of these:

1. free namespace ids when no longer used,
2. reconsider the limit "32" (is it really the most we can fit in our address
   space?),
3. use an exception, not an abort(), if a caller really tries to run more
   concurrent namespaces than we can.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>