namespaces: Create a new per namespace environment library
The purpose of this patch and the following patches is
to give each osv::application the option of having its
own set of environment variables (getenv() and friends).

Some programs, including the OpenMPI framework, do not
use a function like getenv(), but rather directly use
the "extern char **environ" (see environ(7)). For this
reason, the obvious approach of modifying getenv() to
support per-osv::application environment was not enough.
We need each application to see a different "environ"
global variable. We could hack the dynamic linker to
modify the meaning of the "environ" symbol, but this is
ugly. Instead, the solution chosen here was to build a
shared object, libenviron.so, containing "environ",
so that each ELF namespace which loads this library
will get a different environment.

However, this is not the end of the story: Osv's functions,
like getenv(), are linked at build time, and continue to
use the main "environ" when used. We need libenviron.so
to also contain copies of these functions - so when these
copies are used for the first time, the dynamic linker will
resolve their use of "environ" to the one belonging to the
current namespace.

Because of the additional copies of these functions,
libenviron.so adds some memory overhead to every namespace
which uses it (which will be optional). But this overhead is
tiny because these are tiny functions. The beauty of this
approach is that there is no new code needed: libenviron.so
just contains the unmodified "__environ", "getenv", etc. code,
which does exactly the same thing as it does on the main OSv.
The dynamic linker just picks up a different "__environ" object
when running this code.

Note that we compile libenviron.so from the source files,
rather than link the existing object files. The reason is
mundane - usually we compile object files without -fPIC,
but building a shared object requires.

Commit-message-by: Nadav Har'El <nyh@scylladb.com>
Signed-off-by: Benoît Canet <benoit@scylladb.com>
Reviewed-by: Nadav Har'El <nyh@scylladb.com>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>