test: Add test for per namespaces environment variables
As tests/payload-env.so will do a setenv() that will
modify the main OSv environ variable if called without creating
a namespace the tests with and without namespaces are kept
separate as a way to start each test from a clean state.

Still we factorize code by putting common functions in a common
file.

Signed-off-by: Benoît Canet <benoit@scylladb.com>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>