tests: fix incorrect dependency in Makefile
The modules/tests/Makefile used for "scripts/build image=tests" had a
dependency error: We need both the usr.manifest and the actual tests.
But rather than say that both are needed, we only listed usr.manifest,
and said that depended on the test executables. But if the executables
were deleted (by "make clean") and usr.manifest remained, the test
executables would not be rebuilt, and the result is that "scripts/build
image=tests" thinks it built the tests, but it did not, and the upload
of the files into the image will fail.

The solution is to fix the dependencies: the module depends on *both*
usr.manifest and the actual tests. The usr.manifest doesn't really depend
on the test being built (it just lists the tests, making this list doesn't
really depend on the files existing).

"scripts/build image=tests" after "make clean" now works.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>