headers: take inotify.h from newer musl
The include/api/sys/inotify.h we have from old musl is not up-to-date with
LSB 5's declaration of inotify_rm_watch, which says the second parameter
should be an int, not an uint32_t. One of the side-effects is that our
definition in libc/inotify.cc did not become extern "C" (see issue #702).

The solution in this patch is to copy inotify.h from a newer version of
Musl (1.1.11). This is a verbatim copy of that file. The only difference
from the older one we had is in the protoype of inotify_rm_watch().

Fixes #702.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>