test: fail the test on abort
By default, when OSv encounters abort() or a crash, it hangs the VM to
allow a debugger to be attached. This is annoying for automated testing
(using "make check" or "scripts/test.py") because certain bugs in OSv,
the kind which cause a test to crash, hang the test forever instead of
promptly failing it.

The fix in this patch is trivial: Run OSv with the "--power-off-on-abort"
kernel option, which causes abort() to power-off the VM instead of hanging.

The patch also recognizes the string "[backtrace]" (which preceeds a
backtrace on crash) as yet another a sign of test failure, when parsing
the test's output (other messages we looked for, like "Aborted", are
optional, and are not guaranteed to appear).

Fixes #704.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>