tests: fix Boost tests on new distributions
When building and testing OSv on hosts with new Linux distributions based
on gcc 5, such as Fedora 23, all of the Boost-based tests fail, on a
missing symbol. It turns out the problem is that gcc 5 and gcc 4.9 has
different C++ ABIs and the test code we compile on gcc 5 would not be
able to use the Boost library that modules/tests/Makefile takes from
external/ (and was compiled with gcc 4.9).

The "proper" solution would have properly to stop taking the Boost
libraries from external, but doing this would probably cause a whole
cascade of issues. A simpler workaround used in this patch is to use
a magic flag which causes the new libstdc++ header files to revert to
using the old ABI, like they did in the past.

This workaround works, but unfortunately turning on the old ABI now
breaks C++ ABIs which communicated with *new* libraries, like OSv itself.
The only problem this caused in practice is to break OSv-specific "debug()"
calls in some of our Boost-based tests. But this is not a great loss - this
patch simply converts them into std::cerr printouts.

This solution should probably be rethought in the future, but for now,
it works fine, and prevents me from needing to do even bigger changes
to OSv's and OSv tests' build system.

Fixes #686.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>