Makefile: fix libenviron.so compilation error
The "libenviron.so" incorrectly had $^ (the list of dependencies) as the
list of files to compile. This is wrong because on the second build, after
we've built the ".d" dependency files, we now have additional dependencies
which are the header files; These caused us to pass header files in the
compilation command line, resulting in bizarre compilation failures.

This problem manifested as sometimes a "make" would fail stating
compilation errors for the libenviron.so target, but "make clean; make"
would work correctly.

The fix is to use the list of source files, $(environ_sources), not
all the dependencies.

Fixes #708.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>