sched: avoid recursion in timer_list::fired()
This patch fixes a rare OSv crash, a page fault caused by an interrupt-
time recursion which caused the very small interrupt stack to overflow
and page fault (which is not allowed at interrupt time, and therefore
cause a crash).

The problem was as follows: timer_list::fired(), after having run all the
expired timers, calls rearm() to set an interrupt to the next timer coming
up. rearm() calls apic_clock_events::set(timepoint), and if this call
finds that the timepoint has already passed as well, it will run the
callback function immediately, i.e., a recursive call to
timer_list::fired(). In a rare case of many very close timers, this can
lead to deep recursion and a stack overflow.

The solution is to check ourselves whether the next timer has already
passed, and if it has, do not call apic_clock_events::set(timepoint)
but just loop ourselves. When we finally do call apic_clock_events::set(),
call it with a relative time (duration), not a time_point.

We no longer need the apic_clock_events::set(timepoint) variant, and
remove it. This variant just obscures the fact that it needs to call
now() again, and recurses if the given time is in the past.

Signed-off-by: "Nadav Har'El" <nyh@scylladb.com>