elf.cc: avoid overflow of s_obj
Very early in OSv's existance (see commit fa0826a), we added an array
s_obj listing the loaded shared objects, to allow a debugger (via the
"osv syms" command) to find all the loaded objects. However, using a C
array was an over-simplification, and we failed to do any bounds checking
and in issue #711 we came across a case where we overflowed this array.

The solution in this patch is to use std::vector instead of a C array.
The C++ STL takes care for us for all the ugly lookup and deletion code,
and reallocates the array as needed. The code is now simpler, and doesn't
have overflow bugs. What more can we ask for?

Of course, loader.py (where "osv syms" is defined) also needs to be
modified to read the std::vector instead of the C array. But using the
read_vector() function we already have, the new code is even simpler
than the old one.

Fixes #711.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Reviewed-by: Benoît Canet <benoit@scylladb.com>