elf.cc: remove silly "using namespace std"
Using "std::" for standard C++11 stuff is nice and clear, and the accepted
coding standard in OSv, so there was no real reason for the "using namespace
std" in elf.cc. It only saved about 3 uses of "std::" anyway.
And it makes things uglier if we want to use OSv's "mutex" (which becomes
ambiguous, is it std::mutex or OSv's ::mutex?)

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Reviewed-by: Benoît Canet <benoit@scylladb.com>