elf.cc: protect s_objs with mutex
program::s_objs is a static object, shared by all instances of "program"
(a.k.a. ELF namespaces). So we need to protect it against concurrent
modifications, by using a mutex. The chance this would have caused a
bug in practice are slim (when ELF namespaces are used), but not zero.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>