httpserver: also fix usage of getpwuid_r()
The previous patch fixed the usage of getgrgid_r(); But the same bug
existed also in the call to getpwuid_r(), and this patch fixes it.

In both cases, when these functions return 0 ("success") it doesn't mean
the user id or group id was actually found - and we need to check the
returned pointer which is nullptr when the id was not found (and thus
does not have a name).

It is arguable whether our default in this case (write an empty string as
the user name or group name) makes much sense - we should have probably
just written the numeric id - but at least this code will no longer crash.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>