vfs: better handling of failure from filesystem's open()
OSv's VFS open() implementation, sys_open(), uses the desired filesystem's
"open" operation, and should be prepared for this open implementation to
fail. For example, issue #706 is about NFS whose open can fail with EACESS
because the file exists but not allowed to be read by this user (when we
mount an NFS directory, we tell it a specific user to masquerade as).

The problem, discovered by Benoit Canet, is that when opening the file
fails, we call fdrop(fp) on the partially-filled file structure. This is
both unnecessary, and wrong: It is unnecessary because fdrop(fp) was
designed to allow many references to the same file to exist and only the
last fdrop() will cause it to be deleted - while in this case we know
the file has never been opened, and cannot possibly be used by any other
code. Moreover, fdrop(fp) is wrong here because it calls "close()", and
we cannot call close() if open failed (and this probably led the crashes
of issue #706).

So instead, call "delete fp", which is the right thing to do here.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Tested-by: Benoît Canet <benoit@scylladb.com>
Message-Id: <1454247175-14869-1-git-send-email-nyh@scylladb.com>