locale: fix std::toupper(), std::tolower(), and std::regex
OSv's default locale (the "C" locale) correctly supported the character
classes (isprint(), isspace(), etc.) but support for toupper and tolower
was partial and caused breakage of C++ functions which rely on glibc's
implementation. For example, std::regex's constructor used to crash
(as reported in issue #715).

The problem was that the normal C functions toupper() and tolower()
worked, owing to a trivial implementation from musl. However, C++'s
versions of these functions, std::toupper, std::tolower, rely on a
a lookup table which they assume that glibc placed in the locale_t
object. In our default ("C") locale implementation, we put nothing
in that place in locale_t, causing the toupper/tolower attempt to
crash as it uses a lookup table in a random memory location.

This patch adds the missing lookup tables for the default ("C") locale,
which sadly wastes another 2*8*384 = 6144 bytes of OSv memory (I'm not
sure why 2*1*256 = 512 was not deemed enough...). It confirms that now
std::toupper and std::tolower work on the default std::locale, and
also that std::regex's constructor no longer crashes.

Fixes #715

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1454256980-8456-1-git-send-email-nyh@scylladb.com>