published images: drop "monitoring agent"
Since commit 98b88ee3, the "appliance" Capstan images and all pre-prepared
OSv images included a component "monitoring-agent" which reports some
very basic statics (like number of boots) to a central server (on Amazon S3).

This was unnecessary - we've never actually looked at those statistics -
and users started to report it failing. So this patch removes this
module from the image-preparing script, so it will not be included in
the next OSv release.

Note that this module was never automatically included in user-built
images via scripts/build or via Capstan (the Capstan base image did
*not* include this component), so most users never used this module
anyway.

Fixes #714

Signed-off-by: Nadav Har'El <nyh@scylladb.com>