nfs: Discard the new nfs file handle at the end of VOP_CREATE()
VOP_CREATE() is not meant to directly create a reference to the
inode corresponding to the new file.

In the case of sys_open() a subsequent call to VOP_OPEN() will be
done if needed. In the case of mknod() the function does not return
a file descriptor by simply 0 on success.

So discard the nfs file handle right after file creation.

Reported-by: Justin Cinkelj <justin.cinkelj@xlab.si>
Signed-off-by: Benoît Canet <benoit@scylladb.com>
Message-Id: <1455071182-7850-2-git-send-email-benoit@scylladb.com>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>