Remove archaic "/dev/ram0" driver
In the very early days of OSv, when we didn't have a virtio block device
yet, we had a simple "ram disk" (block device in ram) to test filesystem
work. This silly driver remained in OSv, and used up a fixed amount
of memory (4MB) for this ram disk - which on one hand isn't useful, and
on the other hand, a waste of 4 MB.

So this patch drops this driver, and its constant 4 MB memory usage.

Note that this /dev/ram0 ramdisk is *not* the ramfs we use for
holding tools to bootstrap the image (like formatting the ZFS
partition or uploading files). That is a completely separate thing
(ramfs is a read-only file system in RAM, while ram0 is a block device
in RAM)

Fixes #666.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>