scripts/setup.py: add ncurses also to Ubuntu requirements
The modules/libedit module, used in the CLI (in the default built
image), mis-compiles when ncurses is missing on the host (see issue #716).

As a workaround (which should be rethought in the future), tell
scripts/setup.py to install ncurses on the host. We've already been
doing this on Fedora machines, but forgot to do this on Ubuntu as well.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>