__cxa_atexit/finalize: be thread-safe
The Linux C++ ABI uses the functions __cxa_atexit, __cxa_finalize to
implement static constructors in shared objects: When each of the shared
object's global variables is constructed, __cxa_atexit is called to
remember the destructor, and the destructors are finally called when
__cxa_finalze is called.

We had a bug in this implementation: It was not thread-safe, and if
several shared objects were loaded in parallel, and happened to call
their static constructors in parallel, several parallel calls to
__cxa_atexit could lead to a corrupted hash table, and a crash next
time an object is loaded or unloaded and this hash table is consulted
again.

The obvious fix is to use a mutex.

Fixes #723 and fixes #713.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>