Fix reboot loop when abort() happened during early boot.
Our abort() tries to print a backtrace, but cannot do this very early
in the boot process before the dynamic linker is fully set up.
The check for elf::get_program() used to be enough, but because of
recent "ELF namespace" patches, very early in the boot process the
call to elf::get_program() itself crashes, because it calls
sched::thread::current_app() and that assumes there is a current
thread (sched::thread::current() != nullptr). Such a null pointer
dereference before the IDT was fully set up will cause a triple-fault
and hence a reboot - only to have OSv fail in the same way on the
next boot, ad inifintum, which was annoying.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1455700915-27919-1-git-send-email-nyh@scylladb.com>
Signed-off-by: Avi Kivity <avi@scylladb.com>