Fix race of thread::stop_wait() and shutdown
This patch fixes an assertion failure during shutdown that happened
because OSv's thread sleep/wakeup mechanism (and in particular
thread::stop_wait()) can race with the thread::unsafe_stop() mechanism
used during shutdown.

Fixes #729

During shutdown, we call sched::thread::unsafe_stop() on all threads in
an attempt to have them pause to avoid crashes during exit when application
code from mounted filesystems continues to run while the filesystem is
unmounted (see commit 703bf08d72ef617 for more information).

This unsafe_stop() changes threads in the "waiting" state to the
"terminated" state so they will never run again. We can do this with
relative safety for a thread which is "really" waiting; But a thread
preparing to wait might decide, after switching its state to "waiting",
to NOT wait after all because the condition it tests turns out to have
already been fulfilled. In that case, it calls thread::stop_wait().
stop_wait() assumed (and asserts) it will find the thread in either
"waiting" state or some stages of already being woken up - but the
unsafe_stop() code added the possibility that the thread is already in
the "terminated" state, and we need to allow this case as well. In that
case, we can call the scheduler to schedule this thread out (forever).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>