console: delay std::function setting
Since commit e42bf80b7bf, we set the writing std::function once to
avoid creating it (and allocating memory) every time. However, it
turns out that old versions of gcc (such as 4.8) mis-compile this
code, perhaps because we capture "this" in the constructor (?).

So this patch delays setting the std::function to the start()
function. That is also the function which sets _ldisc, and since
we only use the writing function together with _ldisc, it is fine
to set both at the same time.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>